Divided existing plotting functions up into separate subfolder
